Rename DPDK service to HuayuLab NGAF
