+++ services/bin/octobus-tentacles.js
@@ -93,9 +93,9 @@ const services = {
     entryFile: "../ctyun__accessone/bin/ctyun-accessone.js",
     serviceModule: "../ctyun__accessone/src/service.js",
   },
-  "ctdsg-dpdk-v3": {
-    entryFile: "../ctdsg__dpdk_v3/bin/ctdsg-dpdk-v3.js",
-    serviceModule: "../ctdsg__dpdk_v3/src/service.js",
+  "huayulab-ngaf": {
+    entryFile: "../huayulab__ngaf/bin/huayulab-ngaf.js",
+    serviceModule: "../huayulab__ngaf/src/service.js",
   },
   "cloudwalker": {
     entryFile: "../chaitin__cloudwalker/bin/cloudwalker.js",
