+++ services/huayulab__ngaf/package.json
@@ -1,10 +1,10 @@
 {
-  "name": "ctdsg-dpdk-v3",
+  "name": "huayulab-ngaf",
   "version": "0.0.0",
   "private": true,
   "type": "module",
   "bin": {
-    "ctdsg-dpdk-v3": "bin/ctdsg-dpdk-v3.js"
+    "huayulab-ngaf": "bin/huayulab-ngaf.js"
   },
   "dependencies": {
     "@chaitin-ai/octobus-sdk": "^0.5.0"
