+++ services/huayulab__ngaf/secret.schema.json
@@ -10,12 +10,12 @@
     "username": {
       "type": "string",
       "minLength": 1,
-      "description": "CTDSG API username."
+      "description": "HuayuLab NGAF API username."
     },
     "apiSecret": {
       "type": "string",
       "minLength": 1,
-      "description": "CTDSG API secret used to compute md5(md5(apiSecret) + \"-api-!*195\")."
+      "description": "HuayuLab NGAF API secret used to compute md5(md5(apiSecret) + \"-api-!*195\")."
     }
   }
 }
