+++ services/huayulab__ngaf/service.json
@@ -1,8 +1,8 @@
 {
   "schema": "chaitin.octobus.service.v1",
-  "name": "ctdsg-dpdk-v3",
-  "displayName": "CTDSG DPDK V3",
-  "description": "OctoBus package for CTDSG / DPDK 3.0 read-only security operations APIs. Exposes identity, logs, analytics, resource metrics, reference data, and read-only policy object lists through fixed allowlists.",
+  "name": "huayulab-ngaf",
+  "displayName": "HuayuLab NGAF",
+  "description": "OctoBus package for HuayuLab NGAF read-only security operations APIs. Exposes identity, logs, analytics, resource metrics, reference data, and read-only policy object lists through fixed allowlists.",
   "runtime": {
     "mode": "long-running"
   },
@@ -11,7 +11,7 @@
       "proto"
     ],
     "files": [
-      "proto/ctdsg_dpdk_v3.proto"
+      "proto/huayulab_ngaf.proto"
     ]
   },
   "configSchema": "config.schema.json",
@@ -19,39 +19,39 @@
   "sdk": {
     "cli": {
       "commands": {
-        "CTDSG_DPDK_V3.CTDSG_DPDK_V3/GetUserInfo": {
+        "HUAYULAB_NGAF.HUAYULAB_NGAF/GetUserInfo": {
           "name": "get-user-info",
-          "description": "Login to CTDSG DPDK V3 if needed and fetch current API user information."
+          "description": "Login to HuayuLab NGAF if needed and fetch current API user information."
         },
-        "CTDSG_DPDK_V3.CTDSG_DPDK_V3/QuerySecurityLog": {
+        "HUAYULAB_NGAF.HUAYULAB_NGAF/QuerySecurityLog": {
           "name": "query-security-log",
           "description": "Query allowlisted read-only security logs such as IPS, DDoS, AV, WAF, TI, weak-password, outbound, regional, and ICS logs."
         },
-        "CTDSG_DPDK_V3.CTDSG_DPDK_V3/QueryBehaviorLog": {
+        "HUAYULAB_NGAF.HUAYULAB_NGAF/QueryBehaviorLog": {
           "name": "query-behavior-log",
           "description": "Query allowlisted read-only behavior logs such as web, email, IM, login, outbound file, session, alarm, and proxy-bypass logs."
         },
-        "CTDSG_DPDK_V3.CTDSG_DPDK_V3/QueryAuditLog": {
+        "HUAYULAB_NGAF.HUAYULAB_NGAF/QueryAuditLog": {
           "name": "query-audit-log",
           "description": "Query allowlisted read-only protocol, authentication, command, event, and compliance audit logs."
         },
-        "CTDSG_DPDK_V3.CTDSG_DPDK_V3/QuerySecurityStatistic": {
+        "HUAYULAB_NGAF.HUAYULAB_NGAF/QuerySecurityStatistic": {
           "name": "query-security-statistic",
           "description": "Query allowlisted read-only IPS, antivirus, DDoS, and proxy-bypass security statistics."
         },
-        "CTDSG_DPDK_V3.CTDSG_DPDK_V3/QueryFlowAnalysis": {
+        "HUAYULAB_NGAF.HUAYULAB_NGAF/QueryFlowAnalysis": {
           "name": "query-flow-analysis",
           "description": "Query allowlisted read-only traffic and behavior analytics by user, service, site, terminal, location, and online dimensions."
         },
-        "CTDSG_DPDK_V3.CTDSG_DPDK_V3/QueryResourceMetric": {
+        "HUAYULAB_NGAF.HUAYULAB_NGAF/QueryResourceMetric": {
           "name": "query-resource-metric",
           "description": "Query allowlisted read-only CPU, memory, session, online user, and physical interface metrics."
         },
-        "CTDSG_DPDK_V3.CTDSG_DPDK_V3/QueryReferenceData": {
+        "HUAYULAB_NGAF.HUAYULAB_NGAF/QueryReferenceData": {
           "name": "query-reference-data",
-          "description": "Query allowlisted read-only selector and reference data used by DPDK reports."
+          "description": "Query allowlisted read-only selector and reference data used by NGAF reports."
         },
-        "CTDSG_DPDK_V3.CTDSG_DPDK_V3/ListPolicyObjects": {
+        "HUAYULAB_NGAF.HUAYULAB_NGAF/ListPolicyObjects": {
           "name": "list-policy-objects",
           "description": "List allowlisted read-only policy objects such as user blacklist, IP whitelist, and terminal anti-VPN entries."
         }
