+++ services/huayulab__ngaf/src/service.js
@@ -1,8 +1,8 @@
 import { defineService } from "@chaitin-ai/octobus-sdk";
 
-import { handlers } from "./ctdsg-dpdk-v3.js";
+import { handlers } from "./huayulab-ngaf.js";
 
-export { handlers } from "./ctdsg-dpdk-v3.js";
+export { handlers } from "./huayulab-ngaf.js";
 
 export const service = defineService({ handlers });
