+++ services/package.json
@@ -22,7 +22,7 @@
     "cloudatlas": "bin/cloudatlas.js",
     "cosmos": "bin/cosmos.js",
     "ctdsg-fw": "bin/ctdsg-fw.js",
-    "ctdsg-dpdk-v3": "bin/ctdsg-dpdk-v3.js",
+    "huayulab-ngaf": "bin/huayulab-ngaf.js",
     "das-apt": "bin/das-apt.js",
     "cloudwalker": "bin/cloudwalker.js",
     "baiduwaf-waf-web-template": "bin/baiduwaf-waf-web-template.js",
@@ -166,9 +166,9 @@
     "bin/cloudatlas.js",
     "bin/cosmos.js",
     "bin/ctdsg-fw.js",
-    "bin/ctdsg-dpdk-v3.js",
+    "bin/huayulab-ngaf.js",
     "ctdsg__fw",
-    "ctdsg__dpdk_v3",
+    "huayulab__ngaf",
     "bin/das-apt.js",
     "bin/cloudwalker.js",
     "bin/baiduwaf-waf-web-template.js",
